refactor: 함수 모듈로 분리
