+++ public/js/utils.js
@@ -1,3 +1,7 @@
+export function delay(ms) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
 export async function fetchData(url) {
   try {
     const response = await fetch(url);
